+++ examples/text-generation/run_generation.py
@@ -39,7 +39,9 @@ from optimum.habana.checkpoint_utils import (
     model_on_meta,
     write_checkpoints_json,
 )
-from optimum.habana.utils import count_hpu_graphs, get_hpu_memory_stats, override_prints, set_seed, setup_distributed
+from optimum.habana.utils import get_hpu_memory_stats, set_seed
+
+from utils import count_hpu_graphs, override_prints, setup_distributed
 
 
 try:
